rename nominatim.admin to nominatim.tools
